+++ api/database/migrations/2020_10_25_083549_create_skill_category_details_table.php
@@ -16,7 +16,7 @@ class CreateSkillCategoryDetailsTable extends Migration
         Schema::create('skill_category_details', function (Blueprint $table) {
             $table->id();
             $table->string('message', 255)->comment('メッセージ');
-            $table->integer('category_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
             $table->foreign('category_id')->references('id')->on('skill_categories');
             $table->softDeletes();
             $table->timestamps();
